Fix warnings in process-list.cpp (warnings)

+++ src/mt/process-list.cpp
@@ -141,8 +141,8 @@ void System::readProcStat() {
   if (!init_) {
     /// The number of CPU has been detected by going through /proc/stat.
     vCPUData_.resize(cpuNb_ + 1);
-    for (int i = 0; i < (int)cpuNb_; i++)
-      vCPUData_[i].cpu_id_ = i;
+    for (unsigned long i = 0; i < (unsigned long )cpuNb_; i++)
+      vCPUData_[i].cpu_id_ = (int)i;
   }
   aif.close();
 }
